# Conflicts:
#	src/main/java/at/hannibal2/skyhanni/config/features/garden/pests/PestTimerConfig.java

+++ src/main/java/at/hannibal2/skyhanni/config/features/garden/pests/PestTimerConfig.java
@@ -86,7 +86,7 @@ public class PestTimerConfig {
         desc = "Amount of seconds to trigger the pest cooldown warning before cooldown ends."
     )
     @ConfigEditorSlider(minValue = 1, maxValue = 30, minStep = 1)
-    public float cooldownWarningTime = 0;
+    public int cooldownWarningTime = 0;
 
     @Expose
     @ConfigOption(
